Make Region an interface, add MCRegion for vanilla regions
