+++ lib/src/se/llbit/util/Mutable.java
@@ -1,0 +1,17 @@
+package se.llbit.util;
+
+public class Mutable<T> {
+  private T t;
+
+  public Mutable(T t) {
+    this.t = t;
+  }
+
+  public void set(T t) {
+    this.t = t;
+  }
+
+  public T get() {
+    return t;
+  }
+}
